requirements.txt
absl-py==0.8.0
astor==0.8.0
astroid==2.3.1
cycler==0.10.0
gast==0.3.2
google-pasta==0.1.7
grpcio==1.24.0
h5py==2.10.0
imageio==2.5.0
isort==4.3.21
joblib==0.14.0
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
Markdown==3.1.1
matplotlib==3.1.1
mccabe==0.6.1
mock==3.0.5
numpy==1.16.4
Pillow==6.1.0
protobuf==3.9.2
pydot==1.4.1
pylint==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
PyYAML==5.1.2
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
